Minor refactors for new tests and docs (#9564)
